Refactoring package structure
