+++ session-7/main.go
@@ -1,0 +1,8 @@
+package main
+
+import "Module1/methods"
+
+func main() {
+
+	methods.Task4
+}
