--- session-7/model/main.go
@@ -1,1 +1,0 @@
-package model
